Refactor to 8 bit outputs
Truncate to the top 8 bits of the maximum sum, this allows us to attach
outputs directly.

This is not the 100% best, would ideally get only the maximum potential
output from the fir with an input signal instead. For this we may
need to transition to using python, sv or some generator.